backend/requirements.txt
fastapi
uvicorn
anthropic
supabase
python-dotenv
httpx
beautifulsoup4
pydantic
tavily-python
